Warmly send contributions from the master to the develop branch
be precise

+++ .travis-build.sh
@@ -19,6 +19,19 @@ if [ $USER != "travis" ]; then
     exit 1
 fi
 
+if  [ $TRAVIS_SECURE_ENV_VARS = false ] &&
+    [ $TRAVIS_PULL_REQUEST != false ] &&
+    [ $TRAVIS_BRANCH = "master" ]; then
+
+    	printf '=%.0s' {1..70}
+    	printf "\n       し(*･∀･)／   Thanks for the contribution!  ＼(･∀･*)ノ\n"
+    	printf '=%.0s' {1..70}
+    	printf "\n( ＾◡＾)っ Please submit your pull request against the develop branch.\n"
+    	echo   "You can read the contribution guidelines at:"
+    	echo   "https://github.com/syl20bnr/spacemacs/blob/develop/CONTRIBUTING.org"
+    	exit 1
+fi
+
 echo "Pwd $(pwd)"
 rm -rf ~/.emacs.d
 ln -sf `pwd` ~/.emacs.d
